fix: remove invalid nmake syntax from Makefile.frag.w32
The lines attempting to append to PHP_MODULES and all_targets used
GNU Make syntax (variable =  value) which is not supported
by NMAKE. This caused nmake to fail immediately with a syntax error,
preventing the DLL from being built.

Changed to simple 'all: target' declaration which is valid NMAKE syntax.