fix: use npm pack to create platform package tarballs
The build workflow uses `napi artifacts` to create npm package directories,
which are then tarred and uploaded. In the publish workflow, we extract these
tarballs and need to pack each npm subdirectory into a .tgz for publishing.

Previously tried to run `napi artifacts` again, but that expects .node files
in an artifacts directory, not npm package directories.